Move find_crop_size to Trainer
